+++ include/common/monolish_common.hpp
@@ -283,6 +283,33 @@ bool is_same_size(const T &A, const T &B, const types &... args) {
   return is_same_size(A, B) && is_same_size(A, args...);
 }
 
+/**
+/**
+ * @brief compare same device memory status
+ * @note
+ * - # of computation: 1
+ * - Multi-threading: false
+ * - GPU acceleration: false
+ **/
+template <typename T>
+bool is_same_device_mem_stat(const T &arg1, const T &arg2) {
+  return arg1.get_device_mem_stat() == arg2.get_device_mem_stat();
+}
+
+/**
+ * @brief compare same device memory status
+ * @note
+ * - # of computation: 1
+ * - Multi-threading: false
+ * - GPU acceleration: false
+ **/
+template <typename T, typename... types>
+bool is_same_device_mem_stat(const T &arg1, const T &arg2,
+                             const types &... args) {
+  return is_same_device_mem_stat(arg1, arg2) &&
+         is_same_device_mem_stat(arg1, args...);
+}
+
 // create matrix //////////////////
 
 /**