+++ src/internal/monolish_internal.hpp
@@ -1,4 +1,5 @@
 #include <algorithm>
+#include <cassert>
 #include <iostream>
 #include <omp.h>
 #include <stdexcept>