+++ src/vml/double_scalar_vector.cpp
@@ -8,12 +8,8 @@ void vml::add(const vector<double> &a, const double alpha, vector<double> &y) {
   logger.func_in(monolish_func);
 
   // err
-  if (a.size() != y.size()) {
-    throw std::runtime_error("error vector size is not same");
-  }
-  if (a.get_device_mem_stat() != y.get_device_mem_stat()) {
-    throw std::runtime_error("error vector get_device_mem_stat() is not same");
-  }
+  assert(util::is_same_size(a, y));
+  assert(util::is_same_device_mem_stat(a, y));
 
   internal::vadd(y.size(), a.data(), alpha, y.data(), y.get_device_mem_stat());
 
@@ -25,12 +21,8 @@ void vml::sub(const vector<double> &a, const double alpha, vector<double> &y) {
   logger.func_in(monolish_func);
 
   // err
-  if (a.size() != y.size()) {
-    throw std::runtime_error("error vector size is not same");
-  }
-  if (a.get_device_mem_stat() != y.get_device_mem_stat()) {
-    throw std::runtime_error("error vector get_device_mem_stat() is not same");
-  }
+  assert(util::is_same_size(a, y));
+  assert(util::is_same_device_mem_stat(a, y));
 
   internal::vsub(y.size(), a.data(), alpha, y.data(), y.get_device_mem_stat());
 
@@ -42,12 +34,8 @@ void vml::mul(const vector<double> &a, const double alpha, vector<double> &y) {
   logger.func_in(monolish_func);
 
   // err
-  if (a.size() != y.size()) {
-    throw std::runtime_error("error vector size is not same");
-  }
-  if (a.get_device_mem_stat() != y.get_device_mem_stat()) {
-    throw std::runtime_error("error vector get_device_mem_stat() is not same");
-  }
+  assert(util::is_same_size(a, y));
+  assert(util::is_same_device_mem_stat(a, y));
 
   internal::vmul(y.size(), a.data(), alpha, y.data(), y.get_device_mem_stat());
 
@@ -59,12 +47,8 @@ void vml::div(const vector<double> &a, const double alpha, vector<double> &y) {
   logger.func_in(monolish_func);
 
   // err
-  if (a.size() != y.size()) {
-    throw std::runtime_error("error vector size is not same");
-  }
-  if (a.get_device_mem_stat() != y.get_device_mem_stat()) {
-    throw std::runtime_error("error vector get_device_mem_stat() is not same");
-  }
+  assert(util::is_same_size(a, y));
+  assert(util::is_same_device_mem_stat(a, y));
 
   internal::vdiv(y.size(), a.data(), alpha, y.data(), y.get_device_mem_stat());